Update 1
JavaScript syntax is the set of rules, how JavaScript programs are constructed.

JavaScript Programs
A computer program is a list of "instructions" to be "executed" by the computer.

In a programming language, these program instructions are called statements.
JavaScript is a programming language.
JavaScript statements are separated by semicolons:

JavaScript Statements
JavaScript statements are composed of:
Values, Operators, Expressions, Keywords, and Comments.

JavaScript Values
The JavaScript syntax defines two types of values: Fixed values and variable values.

Fixed values are called literals. Variable values are called variables.

JavaScript Literals
The most important rules for writing fixed values are:

Numbers are written with or without decimals:

JavaScript Variables
In a programming language, variables are used to store data values.
JavaScript uses the var keyword to declare variables.
An equal sign is used to assign values to variables.

Not all JavaScript statements are "executed".

Code after double slashes // or between /* and */ is treated as a comment.
Comments are ignored, and will not be executed:

In JavaScript, identifiers are used to name variables (and keywords, and functions, and labels).
The rules for legal names are much the same in most programming languages.
In JavaScript, the first character must be a letter, or an underscore (_), or a dollar sign ($).
Subsequent characters may be letters, digits, underscores, or dollar signs.